agentic_rl_router/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic_rl_router"
version = "0.1.0"
description = "Multi-objective contextual bandit router for Agentic RAG depth/strategy selection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "numpy>=1.26.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentic_rl_router*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
